게시물 수정,삭제시 권한체크

+++ src/main/java/com/sbs/exam/controller/ArticleController.java
@@ -59,6 +59,12 @@ public class ArticleController extends Controller {
     }
     Article article = articleService.getForPrintArticleById(id);
 
+    if (article == null) {
+      rq.historyBack(Util.f("%d번 게시물이 존재하지 않습니다.", id)); // 서식지정자 사용.
+      // rq.historyBack(("%d번 게시물이 존재하지 않습니다.").formatted(id));
+      return;
+    }
+
     rq.setAttr("article", article);
     rq.jsp("../article/detail");
   }
@@ -135,8 +141,15 @@ public class ArticleController extends Controller {
       return;
     }
 
-    ResultData deleteRd = articleService.delete(id); // ResultData(Dto)  보고서 됏다/안됏다를 담음.
-    rq.replace(deleteRd.getMsg(), redirectUri);
+    ResultData actorCanDeleteRd = articleService.actorCanDeleteRd(rq.getLoginedMember(), article);
+
+    if(actorCanDeleteRd.isFail()) {
+      rq.historyBack(actorCanDeleteRd.getMsg());
+      return;
+    }
+
+    articleService.delete(id);
+    rq.replace(Util.f("%d번 게시물을 삭제하였습니다.",id),redirectUri);
 
   }
 
@@ -160,6 +173,13 @@ public class ArticleController extends Controller {
       return;
     }
 
+    ResultData actorCanModifyRd = articleService.actorCanModifyRd(rq.getLoginedMember(), article);
+
+    if(actorCanModifyRd.isFail()) {
+      rq.historyBack(actorCanModifyRd.getMsg());
+      return;
+    }
+
     rq.setAttr("article", article);
     rq.jsp("../article/modify");
   }
@@ -172,7 +192,23 @@ public class ArticleController extends Controller {
 
     if (id == 0) {
       rq.historyBack("id를 입력해주세요.");
+      return;
     }
+
+    Article article = articleService.getForPrintArticleById(id);
+
+    if(article == null) {
+      rq.historyBack(Util.f("%d번 게시물이 존재하지 않습니다.", id));
+      return;
+    }
+
+    ResultData actorCanModifyRd = articleService.actorCanModifyRd(rq.getLoginedMember(), article);
+
+    if(actorCanModifyRd.isFail()) {
+      rq.historyBack(actorCanModifyRd.getMsg());
+      return;
+    }
+
     if (title.length() == 0) {
       rq.historyBack("title을 입력해주세요.");
       return;
